chore: move all cmds to network level
