+++ cli-typescript/src/cmds/mainMenu.ts
@@ -1,26 +1,6 @@
 import { Command } from 'commander';
 import { showMainTitle } from '../utils/display';
 import {
-  createNewWalletCmd,
-  freezeThawContractCmd,
-  checkSignerBalanceCmd,
-  initContractCmd,
-  listProjectsCmd,
-  manageAuthorizedMintersCmd,
-  newProjectCmd,
-  ownerMintCmd,
-  setCosginerCmd,
-  setGlobalWalletLimitCmd,
-  setMaxMintableSupplyCmd,
-  setMintableCmd,
-  setStagesCmd,
-  setTimestampExpiryCmd,
-  setTokenURISuffixCmd,
-  setUriCmd,
-  transferOwnershipCmd,
-  withdrawContractBalanceCmd,
-} from './general';
-import {
   abstract,
   apechain,
   arbitrum,
@@ -43,26 +23,6 @@ export const mainMenu = async () => {
     .name('magicdrop-cli')
     .description('CLI for managing blockchain contracts and tokens')
     .version('2.0.0');
-
-  // Register sub-commands
-  program.addCommand(newProjectCmd);
-  program.addCommand(createNewWalletCmd);
-  program.addCommand(listProjectsCmd);
-  program.addCommand(initContractCmd);
-  program.addCommand(setUriCmd);
-  program.addCommand(setStagesCmd);
-  program.addCommand(setGlobalWalletLimitCmd);
-  program.addCommand(setMaxMintableSupplyCmd);
-  program.addCommand(setCosginerCmd);
-  program.addCommand(setTimestampExpiryCmd);
-  program.addCommand(withdrawContractBalanceCmd);
-  program.addCommand(freezeThawContractCmd);
-  program.addCommand(transferOwnershipCmd);
-  program.addCommand(manageAuthorizedMintersCmd);
-  program.addCommand(setMintableCmd);
-  program.addCommand(setTokenURISuffixCmd);
-  program.addCommand(ownerMintCmd);
-  program.addCommand(checkSignerBalanceCmd);
 
   // network sub-commands
   program.addCommand(eth);
